Refactor: Move users module to src/modules/users with clean controller, domain and usecase structure
